refactor: storage service test code health splits and duplication reduction
