+++ tests/unit/background/services/StorageService.test.js
@@ -1,5 +1,5 @@
 /**
- * StorageService 單元測試
+ * StorageService 單元測試 - 核心行為
  */
 
 import {
@@ -12,12 +12,7 @@ import {
   URL_ALIAS_PREFIX,
   STORAGE_ERROR,
 } from '../../../../scripts/background/services/StorageService.js';
-import {
-  buildDeletedState,
-  buildHighlight,
-  buildSavedPageData,
-  buildStaleStableState,
-} from '../../../helpers/status-fixtures.js';
+import { createStorageServiceHarness } from '../../../helpers/storageServiceTestHarness.js';
 
 jest.mock('../../../../scripts/utils/urlUtils.js', () => {
   const original = jest.requireActual('../../../../scripts/utils/urlUtils.js');
@@ -72,42 +67,13 @@ describe('normalizeUrl', () => {
   });
 });
 
-describe('StorageService', () => {
+describe('StorageService - Core', () => {
   let service = null;
   let mockStorage = null;
   let mockLogger = null;
 
-  function makeStorageMock(data) {
-    mockStorage.local.get.mockImplementation(keys => {
-      const keyList = Array.isArray(keys) ? keys : [keys];
-      return Promise.resolve(
-        Object.fromEntries(keyList.filter(k => k in data).map(k => [k, data[k]]))
-      );
-    });
-  }
-
   beforeEach(() => {
-    mockStorage = {
-      local: {
-        get: jest.fn(() => Promise.resolve({})),
-        set: jest.fn(() => Promise.resolve()),
-        remove: jest.fn(() => Promise.resolve()),
-      },
-      sync: {
-        get: jest.fn(() => Promise.resolve({})),
-        set: jest.fn(() => Promise.resolve()),
-      },
-    };
-    mockLogger = {
-      log: jest.fn(),
-      success: jest.fn(),
-      warn: jest.fn(),
-      error: jest.fn(),
-    };
-    service = new StorageService({
-      chromeStorage: mockStorage,
-      logger: mockLogger,
-    });
+    ({ service, mockStorage, mockLogger } = createStorageServiceHarness(StorageService));
   });
 
   afterEach(() => {
@@ -147,843 +113,10 @@ describe('StorageService', () => {
     });
   });
 
-  describe('getSavedPageData', () => {
-    const setupLegacyReadPath = ({ url, savedData, highlightData = [] }) => {
-      const normalizedUrl = normalizeUrl(url);
-      const pageKey = `${PAGE_PREFIX}${normalizedUrl}`;
-      const savedKey = `${SAVED_PREFIX}${normalizedUrl}`;
-      const aliasKey = `${URL_ALIAS_PREFIX}${normalizedUrl}`;
-      const highlightKey = `${HIGHLIGHTS_PREFIX}${normalizedUrl}`;
-
-      mockStorage.local.get.mockImplementation(keys => {
-        const keyList = Array.isArray(keys) ? keys : [keys];
-        const result = {};
-
-        // _getPageState 讀取：page_* / saved_* / url_alias_*
-        if (keyList.includes(pageKey) && keyList.includes(savedKey) && keyList.includes(aliasKey)) {
-          result[savedKey] = savedData;
-        }
-
-        // _triggerReadTimeUpgrade 讀取：page_* / highlights_*
-        if (keyList.includes(pageKey) && keyList.includes(highlightKey)) {
-          result[highlightKey] = highlightData;
-        }
-
-        return Promise.resolve(result);
-      });
-
-      return { normalizedUrl, pageKey, savedKey, highlightKey };
-    };
-
-    const flushReadTimeUpgrade = async () => {
-      await new Promise(process.nextTick);
-      await new Promise(process.nextTick);
-    };
-
-    it('應該正確獲取保存的頁面數據（page_* 新格式）', async () => {
-      const notionData = { pageId: 'page-123', title: 'Test Page', savedAt: 12_345 };
-      mockStorage.local.get.mockResolvedValue({
-        [`${PAGE_PREFIX}https://example.com/page`]: {
-          notion: notionData,
-          highlights: [],
-          metadata: { createdAt: 12_345, lastUpdated: 12_345 },
-        },
-      });
-
-      const result = await service.getSavedPageData('https://example.com/page');
-      // Phase 3: 返回映射後的 notion 子欄位 (notionPageId, notionUrl, etc.)
-      expect(result).toEqual({
-        notionPageId: 'page-123',
-        notionUrl: null,
-        title: 'Test Page',
-        savedAt: 12_345,
-        lastVerifiedAt: null,
-        destinationProfileId: null,
-      });
-    });
-
-    it('應該在沒有數據時返回 null', async () => {
-      const result = await service.getSavedPageData('https://example.com/page');
-      expect(result).toBeNull();
-    });
-
-    it('storage.local.get 回傳 null 時應視為空結果', async () => {
-      mockStorage.local.get.mockResolvedValue(null);
-
-      const result = await service.getSavedPageData('https://example.com/page');
-
-      expect(result).toBeNull();
-    });
-
-    it('應該在 page_* 的 notion 為 null 時返回 null', async () => {
-      mockStorage.local.get.mockResolvedValue({
-        [`${PAGE_PREFIX}https://example.com/page`]: {
-          notion: null,
-          highlights: [{ id: 'h1', text: 'test' }],
-          metadata: { createdAt: 12_345, lastUpdated: 12_345 },
-        },
-      });
-
-      const result = await service.getSavedPageData('https://example.com/page');
-      expect(result).toBeNull();
-    });
-
-    it('應在直接查找失敗時嘗試使用 alias 查找（page_* 格式）', async () => {
-      const originalUrl = 'https://example.com/original';
-      const stableUrl = 'https://example.com/stable';
-      const notionData = { pageId: 'page-abc', title: 'Test Page' };
-
-      mockStorage.local.get.mockImplementation(keys => {
-        const result = {};
-        const keyList = Array.isArray(keys) ? keys : [keys];
-
-        keyList.forEach(k => {
-          if (k === `${URL_ALIAS_PREFIX}${originalUrl}`) {
-            result[k] = stableUrl;
-          }
-          if (k === `${PAGE_PREFIX}${stableUrl}`) {
-            result[k] = { notion: notionData, highlights: [], metadata: {} };
-          }
-        });
-
-        return Promise.resolve(result);
-      });
-
-      const result = await service.getSavedPageData(originalUrl);
-
-      // Phase 3: 返回映射後的 notion 子欄位
-      expect(result).toEqual({
-        notionPageId: 'page-abc',
-        notionUrl: null,
-        title: 'Test Page',
-        savedAt: null,
-        lastVerifiedAt: null,
-        destinationProfileId: null,
-      });
-    });
-
-    it('缺少 alias 時，不應把孤立的 stable page_* 視為 original URL 的已保存狀態', async () => {
-      const originalUrl = 'https://example.com/original';
-      const stableUrl = 'https://example.com/stable';
-      const storageData = buildStaleStableState({
-        originalUrl,
-        stableUrl,
-        includeAlias: false,
-      });
-
-      mockStorage.local.get.mockImplementation(keys => {
-        const keyList = Array.isArray(keys) ? keys : [keys];
-        return Promise.resolve(
-          Object.fromEntries(
-            keyList.filter(key => key in storageData).map(key => [key, storageData[key]])
-          )
-        );
-      });
-
-      const result = await service.getSavedPageData(originalUrl);
-
-      expect(result).toBeNull();
-    });
-
-    it('alias 指向 notion:null 的 stable page_* 時，應回傳 null 而非已保存', async () => {
-      const originalUrl = 'https://example.com/original';
-      const stableUrl = 'https://example.com/stable';
-      const storageData = buildDeletedState({
-        originalUrl,
-        stableUrl,
-        highlights: [buildHighlight()],
-      });
-
-      mockStorage.local.get.mockImplementation(keys => {
-        const keyList = Array.isArray(keys) ? keys : [keys];
-        return Promise.resolve(
-          Object.fromEntries(
-            keyList.filter(key => key in storageData).map(key => [key, storageData[key]])
-          )
-        );
-      });
-
-      const result = await service.getSavedPageData(originalUrl);
-
-      expect(result).toBeNull();
-    });
-
-    it('讀時升級第一次失敗後，應記錄重試狀態與 nextRetryAt', async () => {
-      const url = 'https://example.com/retry-first-failure';
-      const savedData = { notionPageId: 'legacy-1', title: 'legacy title' };
-      const { normalizedUrl } = setupLegacyReadPath({ url, savedData });
-
-      jest.spyOn(Date, 'now').mockReturnValue(1_700_000_000_000);
-      jest.spyOn(Math, 'random').mockReturnValue(0);
-      mockStorage.local.set.mockRejectedValue(new Error('upgrade write failed'));
-
-      const result = await service.getSavedPageData(url);
-      await flushReadTimeUpgrade();
-
-      expect(result).toEqual(savedData);
-      const retryState = service._failedUpgradeAttempts.get(normalizedUrl);
-      expect(retryState).toEqual({
-        attempts: 1,
-        firstFailureAt: 1_700_000_000_000,
-        lastFailureAt: 1_700_000_000_000,
-        nextRetryAt: 1_700_000_000_500,
-      });
-    });
-
-    it('尚未到達 nextRetryAt 時，應跳過重試', async () => {
-      const url = 'https://example.com/retry-skip-before-next-window';
-      const savedData = { notionPageId: 'legacy-2', title: 'legacy title' };
-      const { normalizedUrl } = setupLegacyReadPath({ url, savedData });
-
-      const now = 1_700_000_001_000;
-      jest.spyOn(Date, 'now').mockReturnValue(now);
-      service._failedUpgradeAttempts.set(normalizedUrl, {
-        attempts: 1,
-        firstFailureAt: now - 1000,
-        lastFailureAt: now - 1000,
-        nextRetryAt: now + 30_000,
-      });
-
-      const result = await service.getSavedPageData(url);
-      await flushReadTimeUpgrade();
-
-      expect(result).toEqual(savedData);
-      expect(mockStorage.local.set).not.toHaveBeenCalled();
-      expect(mockStorage.local.remove).not.toHaveBeenCalled();
-      expect(service._failedUpgradeAttempts.get(normalizedUrl)).toEqual(
-        expect.objectContaining({
-          attempts: 1,
-          nextRetryAt: now + 30_000,
-        })
-      );
-    });
-
-    it('超過 nextRetryAt 後，應允許再次嘗試讀時升級', async () => {
-      const url = 'https://example.com/retry-after-next-window';
-      const savedData = { notionPageId: 'legacy-3', title: 'legacy title' };
-      const { normalizedUrl } = setupLegacyReadPath({ url, savedData });
-
-      const now = 1_700_000_002_000;
-      jest.spyOn(Date, 'now').mockReturnValue(now);
-      jest.spyOn(Math, 'random').mockReturnValue(0);
-      service._failedUpgradeAttempts.set(normalizedUrl, {
-        attempts: 1,
-        firstFailureAt: now - 2000,
-        lastFailureAt: now - 1500,
-        nextRetryAt: now - 1,
-      });
-      mockStorage.local.set.mockRejectedValue(new Error('upgrade write failed again'));
-
-      const result = await service.getSavedPageData(url);
-      await flushReadTimeUpgrade();
-
-      expect(result).toEqual(savedData);
-      expect(mockStorage.local.set).toHaveBeenCalledTimes(1);
-      expect(service._failedUpgradeAttempts.get(normalizedUrl)).toEqual({
-        attempts: 2,
-        firstFailureAt: now - 2000,
-        lastFailureAt: now,
-        nextRetryAt: now + 1000,
-      });
-    });
-
-    it('讀時升級成功後，應清除失敗追蹤狀態', async () => {
-      const url = 'https://example.com/retry-clear-on-success';
-      const savedData = { notionPageId: 'legacy-4', title: 'legacy title' };
-      const { normalizedUrl } = setupLegacyReadPath({ url, savedData });
-
-      const now = 1_700_000_003_000;
-      jest.spyOn(Date, 'now').mockReturnValue(now);
-      service._failedUpgradeAttempts.set(normalizedUrl, {
-        attempts: 2,
-        firstFailureAt: now - 5000,
-        lastFailureAt: now - 1000,
-        nextRetryAt: now - 1,
-      });
-      mockStorage.local.set.mockResolvedValue();
-      mockStorage.local.remove.mockResolvedValue();
-
-      const result = await service.getSavedPageData(url);
-      await flushReadTimeUpgrade();
-
-      expect(result).toEqual(savedData);
-      expect(mockStorage.local.set).toHaveBeenCalledTimes(1);
-      expect(mockStorage.local.remove).toHaveBeenCalledTimes(1);
-      expect(service._failedUpgradeAttempts.has(normalizedUrl)).toBe(false);
-    });
-
-    it('達到 maxAttempts 且仍在 TTL 內時，應跳過重試', async () => {
-      const url = 'https://example.com/retry-max-attempts';
-      const savedData = { notionPageId: 'legacy-5', title: 'legacy title' };
-      const { normalizedUrl } = setupLegacyReadPath({ url, savedData });
-
-      const now = 1_700_000_004_000;
-      jest.spyOn(Date, 'now').mockReturnValue(now);
-      service._failedUpgradeAttempts.set(normalizedUrl, {
-        attempts: 5,
-        firstFailureAt: now - 10_000, // 仍在 30 分鐘 TTL 內
-        lastFailureAt: now - 1000,
-        nextRetryAt: now - 1,
-      });
-
-      const result = await service.getSavedPageData(url);
-      await flushReadTimeUpgrade();
-
-      expect(result).toEqual(savedData);
-      expect(mockStorage.local.set).not.toHaveBeenCalled();
-      expect(mockStorage.local.remove).not.toHaveBeenCalled();
-      expect(service._failedUpgradeAttempts.get(normalizedUrl)).toEqual(
-        expect.objectContaining({
-          attempts: 5,
-        })
-      );
-    });
-
-    it('應在第 5 次失敗時將 nextRetryAt 設為 firstFailureAt + TTL', async () => {
-      const url = 'https://example.com/retry-hit-max-attempt';
-      const savedData = { notionPageId: 'legacy-7', title: 'legacy title' };
-      const { normalizedUrl } = setupLegacyReadPath({ url, savedData });
-
-      const now = 1_700_000_004_500;
-      const ttlMs = 30 * 60 * 1000;
-      const firstFailureAt = now - 20_000;
-      jest.spyOn(Date, 'now').mockReturnValue(now);
-
-      service._failedUpgradeAttempts.set(normalizedUrl, {
-        attempts: 4,
-        firstFailureAt,
-        lastFailureAt: now - 1000,
-        nextRetryAt: now - 1,
-      });
-      mockStorage.local.set.mockRejectedValue(new Error('upgrade write failed at attempt 5'));
-
-      const result = await service.getSavedPageData(url);
-      await flushReadTimeUpgrade();
-
-      expect(result).toEqual(savedData);
-      expect(mockStorage.local.set).toHaveBeenCalledTimes(1);
-      expect(service._failedUpgradeAttempts.get(normalizedUrl)).toEqual({
-        attempts: 5,
-        firstFailureAt,
-        lastFailureAt: now,
-        nextRetryAt: firstFailureAt + ttlMs,
-      });
-    });
-
-    it('超過 TTL 後，應重置並允許新一輪嘗試', async () => {
-      const url = 'https://example.com/retry-reset-after-ttl';
-      const savedData = { notionPageId: 'legacy-6', title: 'legacy title' };
-      const { normalizedUrl } = setupLegacyReadPath({ url, savedData });
-
-      const now = 1_700_000_005_000;
-      const ttlMs = 30 * 60 * 1000;
-      jest.spyOn(Date, 'now').mockReturnValue(now);
-      jest.spyOn(Math, 'random').mockReturnValue(0);
-      service._failedUpgradeAttempts.set(normalizedUrl, {
-        attempts: 5,
-        firstFailureAt: now - ttlMs - 1,
-        lastFailureAt: now - ttlMs - 1,
-        nextRetryAt: now + 999_999,
-      });
-      mockStorage.local.set.mockRejectedValue(new Error('upgrade write failed after ttl'));
-
-      const result = await service.getSavedPageData(url);
-      await flushReadTimeUpgrade();
-
-      expect(result).toEqual(savedData);
-      expect(mockStorage.local.set).toHaveBeenCalledTimes(1);
-      expect(service._failedUpgradeAttempts.get(normalizedUrl)).toEqual({
-        attempts: 1,
-        firstFailureAt: now,
-        lastFailureAt: now,
-        nextRetryAt: now + 500,
-      });
-    });
-  });
-
-  describe('setSavedPageData destination profile metadata', () => {
-    it('應寫入 page_* 的 notion.destinationProfileId', async () => {
-      const url = 'https://example.com/page';
-
-      await service.setSavedPageData(url, {
-        notionPageId: 'page-1',
-        notionUrl: 'https://www.notion.so/page-1',
-        title: 'Saved Page',
-        destinationProfileId: 'pid',
-      });
-
-      expect(mockStorage.local.set).toHaveBeenCalledWith({
-        [`${PAGE_PREFIX}${url}`]: expect.objectContaining({
-          notion: expect.objectContaining({
-            pageId: 'page-1',
-            destinationProfileId: 'pid',
-          }),
-        }),
-      });
-    });
-
-    it('destinationProfileId 傳入 null 時應清除既有目的地欄位', async () => {
-      const url = 'https://example.com/page';
-      makeStorageMock({
-        [`${PAGE_PREFIX}${url}`]: {
-          notion: {
-            pageId: 'page-1',
-            title: 'Saved Page',
-            destinationProfileId: 'pid',
-          },
-          highlights: [],
-          metadata: {},
-        },
-      });
-
-      await service.setSavedPageData(url, {
-        notionPageId: 'page-1',
-        destinationProfileId: null,
-      });
-
-      expect(mockStorage.local.set).toHaveBeenCalledWith({
-        [`${PAGE_PREFIX}${url}`]: expect.objectContaining({
-          notion: expect.objectContaining({
-            destinationProfileId: null,
-          }),
-        }),
-      });
-    });
-
-    it('直接 page_* 查找失敗但 alias 存在時，set/get 應使用 alias stable URL', async () => {
-      const originalUrl = 'https://example.com/original';
-      const stableUrl = 'https://example.com/stable';
-      const storageData = {
-        [`${URL_ALIAS_PREFIX}${originalUrl}`]: stableUrl,
-        [`${PAGE_PREFIX}${stableUrl}`]: {
-          notion: {
-            pageId: 'page-1',
-            title: 'Stable Page',
-            destinationProfileId: null,
-          },
-          highlights: [{ id: 'h1', text: 'highlight' }],
-          metadata: {},
-        },
-      };
-      mockStorage.local.get.mockImplementation(keys => {
-        const keyList = Array.isArray(keys) ? keys : [keys];
-        return Promise.resolve(
-          Object.fromEntries(
-            keyList.filter(key => key in storageData).map(key => [key, storageData[key]])
-          )
-        );
-      });
-
-      await expect(service.getSavedPageData(originalUrl)).resolves.toEqual(
-        expect.objectContaining({ notionPageId: 'page-1' })
-      );
-      await service.setSavedPageData(originalUrl, {
-        notionPageId: 'page-1',
-        destinationProfileId: 'pid',
-      });
-
-      expect(mockStorage.local.set).toHaveBeenCalledWith({
-        [`${PAGE_PREFIX}${stableUrl}`]: expect.objectContaining({
-          notion: expect.objectContaining({
-            destinationProfileId: 'pid',
-          }),
-        }),
-      });
-    });
-
-    it('alias 命中 legacy stable saved_* 時，set 應寫入 canonical stable page_* key', async () => {
-      const originalUrl = 'https://example.com/original';
-      const stableUrl = 'https://example.com/stable';
-      const storageData = {
-        [`${URL_ALIAS_PREFIX}${originalUrl}`]: stableUrl,
-        [`${SAVED_PREFIX}${stableUrl}`]: {
-          notionPageId: 'page-legacy',
-          notionUrl: 'https://www.notion.so/page-legacy',
-          title: 'Legacy Stable Page',
-          savedAt: 1,
-        },
-        [`${HIGHLIGHTS_PREFIX}${stableUrl}`]: [{ id: 'h1', text: 'highlight' }],
-      };
-      mockStorage.local.get.mockImplementation(keys => {
-        const keyList = Array.isArray(keys) ? keys : [keys];
-        return Promise.resolve(
-          Object.fromEntries(
-            keyList.filter(key => key in storageData).map(key => [key, storageData[key]])
-          )
-        );
-      });
-
-      await service.setSavedPageData(originalUrl, {
-        notionPageId: 'page-legacy',
-        destinationProfileId: 'pid',
-      });
-
-      expect(mockStorage.local.set).toHaveBeenCalledWith({
-        [`${PAGE_PREFIX}${stableUrl}`]: expect.objectContaining({
-          notion: expect.objectContaining({
-            pageId: 'page-legacy',
-            destinationProfileId: 'pid',
-          }),
-        }),
-      });
-      expect(mockStorage.local.set).not.toHaveBeenCalledWith(
-        expect.objectContaining({
-          [`${PAGE_PREFIX}${originalUrl}`]: expect.anything(),
-        })
-      );
-      expect(mockStorage.local.remove).toHaveBeenCalledWith(
-        expect.arrayContaining([`${SAVED_PREFIX}${stableUrl}`])
-      );
-    });
-
-    it('[CANONICAL] alias 命中且僅 page_<original> 有殘留時 MUST 寫入 page_<stableUrl> 並清掉 page_<original>', async () => {
-      const originalUrl = 'https://example.com/article?utm=abc';
-      const stableUrl = 'https://example.com/article';
-      const aliasNormKey = `${URL_ALIAS_PREFIX}${originalUrl}`;
-      const stablePageKey = `${PAGE_PREFIX}${stableUrl}`;
-      const originalPageKey = `${PAGE_PREFIX}${originalUrl}`;
-
-      const storageData = {
-        [aliasNormKey]: stableUrl,
-        [originalPageKey]: {
-          notion: { pageId: 'old', destinationProfileId: null },
-          highlights: [{ id: 'h-stale', text: 'stale' }],
-          metadata: { lastUpdated: 100 },
-        },
-      };
-      mockStorage.local.get.mockImplementation(keys => {
-        const list = Array.isArray(keys) ? keys : [keys];
-        return Promise.resolve(
-          Object.fromEntries(list.filter(k => k in storageData).map(k => [k, storageData[k]]))
-        );
-      });
-
-      await service.setSavedPageData(originalUrl, {
-        notionPageId: 'page-canonical',
-        destinationProfileId: 'pid-1',
-      });
-
-      // 寫入應落在 canonical stable page key
-      expect(mockStorage.local.set).toHaveBeenCalledWith({
-        [stablePageKey]: expect.objectContaining({
-          notion: expect.objectContaining({
-            pageId: 'page-canonical',
-            destinationProfileId: 'pid-1',
-          }),
-        }),
-      });
-      // MUST NOT 寫回 page_<original>
-      expect(mockStorage.local.set).not.toHaveBeenCalledWith(
-        expect.objectContaining({ [originalPageKey]: expect.anything() })
-      );
-      // legacy 殘留 page_<original> MUST 被清掉
-      const removedKeys = mockStorage.local.remove.mock.calls.flatMap(args => args[0]);
-      expect(removedKeys).toEqual(expect.arrayContaining([originalPageKey]));
-    });
-
-    it('destinationProfileId 明確傳入 undefined 時應寫入 null', async () => {
-      const url = 'https://example.com/page';
-
-      await service.setSavedPageData(url, {
-        notionPageId: 'page-1',
-        destinationProfileId: undefined,
-      });
-
-      expect(mockStorage.local.set).toHaveBeenCalledWith({
-        [`${PAGE_PREFIX}${url}`]: expect.objectContaining({
-          notion: expect.objectContaining({
-            destinationProfileId: null,
-          }),
-        }),
-      });
-    });
-  });
-
-  describe('getHighlights', () => {
-    const NORM_URL = 'https://example.com/posts/hello';
-    const RAW_URL = 'https://example.com/posts/hello?utm_source=twitter#frag';
-    const STABLE_URL = 'https://example.com/?p=123';
-    const PAGE_NORM = `${PAGE_PREFIX}${NORM_URL}`;
-    const PAGE_STABLE = `${PAGE_PREFIX}${STABLE_URL}`;
-    const HL_NORM = `${HIGHLIGHTS_PREFIX}${NORM_URL}`;
-    const ALIAS_NORM = `${URL_ALIAS_PREFIX}${NORM_URL}`;
-    const ALIAS_RAW = `${URL_ALIAS_PREFIX}${RAW_URL}`;
-    const SAMPLE_HIGHLIGHTS = [{ id: 'h1', text: 'Hello', color: 'yellow' }];
-
-    it('無 alias：命中 page_<normalizedUrl>，回傳 highlights', async () => {
-      makeStorageMock({
-        [PAGE_NORM]: { highlights: SAMPLE_HIGHLIGHTS, notion: null },
-      });
-      const result = await service.getHighlights(NORM_URL);
-      expect(result).toEqual(SAMPLE_HIGHLIGHTS);
-    });
-
-    it('無 alias：page_* 不存在，fallback 到 highlights_<normalizedUrl>（純陣列）', async () => {
-      makeStorageMock({ [HL_NORM]: SAMPLE_HIGHLIGHTS });
-      const result = await service.getHighlights(NORM_URL);
-      expect(result).toEqual(SAMPLE_HIGHLIGHTS);
-    });
-
-    it('無 alias：所有 key 均不存在，回傳 null', async () => {
-      makeStorageMock({});
-      const result = await service.getHighlights(NORM_URL);
-      expect(result).toBeNull();
-    });
-
-    it('有 alias：alias 指向 STABLE_URL，優先命中 page_<stableUrl>', async () => {
-      makeStorageMock({
-        [ALIAS_NORM]: STABLE_URL,
-        [PAGE_STABLE]: { highlights: SAMPLE_HIGHLIGHTS, notion: { pageId: 'abc' } },
-        [PAGE_NORM]: { highlights: [{ id: 'stale' }], notion: null },
-      });
-      const result = await service.getHighlights(NORM_URL);
-      expect(result).toEqual(SAMPLE_HIGHLIGHTS);
-    });
-
-    it('有 alias：page_<stableUrl> miss，fallback 到 page_<normalizedUrl>', async () => {
-      const normHighlights = [{ id: 'hn', text: 'norm', color: 'green' }];
-      makeStorageMock({
-        [ALIAS_NORM]: STABLE_URL,
-        // PAGE_STABLE 不存在
-        [PAGE_NORM]: { highlights: normHighlights, notion: null },
-      });
-      const result = await service.getHighlights(NORM_URL);
-      expect(result).toEqual(normHighlights);
-    });
-
-    it('有 alias 但非法（非 http）：退回無 alias 路徑，命中 page_<normalizedUrl>', async () => {
-      makeStorageMock({
-        [ALIAS_NORM]: 'ftp://example.com/file', // 非法 alias
-        [PAGE_NORM]: { highlights: SAMPLE_HIGHLIGHTS, notion: null },
-      });
-      const result = await service.getHighlights(NORM_URL);
-      expect(result).toEqual(SAMPLE_HIGHLIGHTS);
-    });
-
-    it('page_* highlights 欄位損壞（非陣列）：fallback 到 highlights_*', async () => {
-      makeStorageMock({
-        [PAGE_NORM]: { highlights: 'corrupted', notion: null },
-        [HL_NORM]: SAMPLE_HIGHLIGHTS,
-      });
-      const result = await service.getHighlights(NORM_URL);
-      expect(result).toEqual(SAMPLE_HIGHLIGHTS);
-    });
-
-    it('有 alias：page_* miss，fallback 到 highlights_<stableUrl>（alias 舊格式）', async () => {
-      const HL_STABLE = `${HIGHLIGHTS_PREFIX}${STABLE_URL}`;
-      const stableHlHighlights = [{ id: 'hs', text: 'stable legacy', color: 'blue' }];
-      makeStorageMock({
-        [ALIAS_NORM]: STABLE_URL,
-        // PAGE_STABLE 和 PAGE_NORM 均不存在
-        [HL_STABLE]: stableHlHighlights, // alias-resolved 舊格式 key
-        [HL_NORM]: [{ id: 'hn', text: 'norm legacy', color: 'red' }],
-      });
-      const result = await service.getHighlights(NORM_URL);
-      // 應優先命中 highlights_<stableUrl>，而非 highlights_<normalizedUrl>
-      expect(result).toEqual(stableHlHighlights);
-    });
-
-    it('[REGRESSION] rawUrl alias key 存在時，應能透過 alias 命中 page_<stableUrl>', async () => {
-      makeStorageMock({
-        [ALIAS_RAW]: STABLE_URL,
-        [PAGE_STABLE]: { highlights: SAMPLE_HIGHLIGHTS, notion: { pageId: 'abc' } },
-      });
-
-      const result = await service.getHighlights(RAW_URL);
-
-      expect(result).toEqual(SAMPLE_HIGHLIGHTS);
-      expect(mockStorage.local.get).toHaveBeenCalledWith(
-        expect.arrayContaining([
-          ALIAS_NORM,
-          ALIAS_RAW,
-          PAGE_NORM,
-          `${PAGE_PREFIX}${RAW_URL}`,
-          HL_NORM,
-          `${HIGHLIGHTS_PREFIX}${RAW_URL}`,
-        ])
-      );
-    });
-
-    it('storage 不可用時應拋出 STORAGE_ERROR', async () => {
-      // 直接覆寫 storage 為 null（Jest 環境有 chrome global，constructor 會 fallback）
-      service.storage = null;
-      await expect(service.getHighlights(NORM_URL)).rejects.toThrow(STORAGE_ERROR);
-    });
-  });
-
-  describe('setSavedPageData', () => {
-    it('應該正確設置頁面數據（Phase 3: page_*.notion partial update）', async () => {
-      // 先設定無既有資料（新建路徑）
-      mockStorage.local.get.mockResolvedValue({});
-      const data = { title: 'Test Page', pageId: 'page-123' };
-      await service.setSavedPageData('https://example.com/page', data);
-
-      expect(mockStorage.local.set).toHaveBeenCalledWith(
-        expect.objectContaining({
-          [`${PAGE_PREFIX}https://example.com/page`]: expect.objectContaining({
-            notion: expect.objectContaining({ title: 'Test Page' }),
-            metadata: expect.objectContaining({ lastUpdated: expect.any(Number) }),
-          }),
-        })
-      );
-    });
-
-    it('應該在刪除舊 saved_* key 失敗時記錄 debug 日誌', async () => {
-      mockStorage.local.get.mockResolvedValue({});
-      mockStorage.local.remove.mockRejectedValue(new Error('Remove failed'));
-      // 使用 optional chaining 的 mock
-      mockLogger.debug = jest.fn();
-
-      const data = { title: 'Test Page', pageId: 'page-123' };
-      await service.setSavedPageData('https://example.com/page', data);
-
-      // Verify that data was still set correctly despite the remove error
-      expect(mockStorage.local.set).toHaveBeenCalled();
-
-      // Since remove is non-blocking (Promise.catch), we might need to wait for the microtask queue
-      await new Promise(process.nextTick);
-
-      expect(mockLogger.debug).toHaveBeenCalledWith(
-        '[StorageService] Failed to remove legacy keys',
-        expect.objectContaining({ error: 'Remove failed' })
-      );
-    });
-
-    it('應該在 page_* 不存在時，從舊格式 highlights_* 取回並保留標注', async () => {
-      const pageKey = `${PAGE_PREFIX}https://example.com/page`;
-      const hlKey = `${HIGHLIGHTS_PREFIX}https://example.com/page`;
-      const legacyHighlights = [
-        { id: 'h1', text: 'first highlight', color: 'yellow' },
-        { id: 'h2', text: 'second highlight', color: 'green' },
-      ];
-
-      // page_* 不存在，highlights_* 存在（舊格式純陣列）
-      mockStorage.local.get.mockResolvedValue({ [hlKey]: legacyHighlights });
-
-      const data = { title: 'Test Page', notionPageId: 'page-abc', savedAt: 1000 };
-      await service.setSavedPageData('https://example.com/page', data);
-
-      const setPayload = mockStorage.local.set.mock.calls[0][0][pageKey];
-      expect(setPayload.notion).toEqual(expect.objectContaining({ pageId: 'page-abc' }));
-      expect(setPayload.highlights).toEqual([
-        expect.objectContaining({
-          id: 'h1',
-          text: 'first highlight',
-          color: 'yellow',
-          rangeInfo: expect.any(Object),
-          timestamp: expect.any(Number),
-        }),
-        expect.objectContaining({
-          id: 'h2',
-          text: 'second highlight',
-          color: 'green',
-          rangeInfo: expect.any(Object),
-          timestamp: expect.any(Number),
-        }),
-      ]);
-      // highlights_* 應被清理（已遷移到 page_*）
-      expect(mockStorage.local.remove).toHaveBeenCalledWith(expect.arrayContaining([hlKey]));
-    });
-
-    it('應該在 page_* 已存在時，保留其 highlights 而非回退到 highlights_*', async () => {
-      const pageKey = `${PAGE_PREFIX}https://example.com/page`;
-      const hlKey = `${HIGHLIGHTS_PREFIX}https://example.com/page`;
-      const existingHighlights = [
-        {
-          id: 'h3',
-          text: 'existing in page_*',
-          color: 'yellow',
-          rangeInfo: { start: 0, end: 3 },
-          timestamp: 1_700_000_000_001,
-        },
-      ];
-      const staleHighlights = [{ id: 'h1', text: 'stale in highlights_*' }];
-
-      // page_* 已存在（有自己的 highlights），highlights_* 也存在（舊資料）
-      mockStorage.local.get.mockResolvedValue({
-        [pageKey]: { highlights: existingHighlights, notion: null, metadata: {} },
-        [hlKey]: staleHighlights,
-      });
-
-      const data = { title: 'Test', notionPageId: 'page-xyz' };
-      await service.setSavedPageData('https://example.com/page', data);
-
-      // 應使用 page_* 的 highlights，而非 highlights_* 的舊資料
-      expect(mockStorage.local.set).toHaveBeenCalledWith(
-        expect.objectContaining({
-          [pageKey]: expect.objectContaining({
-            highlights: existingHighlights,
-          }),
-        })
-      );
-    });
-
-    it('應該在 current.highlights 非陣列時覆寫為合法 highlights 陣列', async () => {
-      const pageKey = `${PAGE_PREFIX}https://example.com/page`;
-      const hlKey = `${HIGHLIGHTS_PREFIX}https://example.com/page`;
-      mockStorage.local.get.mockResolvedValue({
-        [pageKey]: { highlights: { invalid: true }, notion: null, metadata: {} },
-        [hlKey]: [{ id: 'legacy', text: 'legacy text', color: 'yellow' }],
-      });
-
-      const data = { title: 'Test', notionPageId: 'page-xyz' };
-      await service.setSavedPageData('https://example.com/page', data);
-
-      const setPayload = mockStorage.local.set.mock.calls[0][0][pageKey];
-      expect(Array.isArray(setPayload.highlights)).toBe(true);
-      expect(setPayload.highlights).toEqual([]);
-    });
-
-    it('寫入 { lastVerifiedAt: 0, savedAt: 0 } 等 falsy 值後可原值保留，不被轉為 null', async () => {
-      mockStorage.local.get.mockResolvedValue({});
-      const data = { title: 'Test Page', pageId: 'page-123', lastVerifiedAt: 0, savedAt: 0 };
-      await service.setSavedPageData('https://example.com/page', data);
-
-      const setPayload =
-        mockStorage.local.set.mock.calls[0][0][`${PAGE_PREFIX}https://example.com/page`];
-      expect(setPayload.notion.lastVerifiedAt).toBe(0);
-      expect(setPayload.notion.savedAt).toBe(0);
-    });
-
-    it('遇到空字串 notion 欄位時應保留現有有效 pageId 與 url', async () => {
-      const pageKey = `${PAGE_PREFIX}https://example.com/page`;
-      mockStorage.local.get.mockResolvedValue({
-        [pageKey]: {
-          highlights: [],
-          notion: {
-            pageId: 'existing-page-id',
-            url: 'https://www.notion.so/existing-page-id',
-            title: 'Existing Title',
-            savedAt: 123,
-            lastVerifiedAt: 456,
-          },
-          metadata: {},
-        },
-      });
-
-      const data = {
-        title: 'Updated Title',
-        notionPageId: '',
-        pageId: '   ',
-        notionUrl: '',
-        url: '   ',
-      };
-
-      await service.setSavedPageData('https://example.com/page', data);
-
-      const setPayload = mockStorage.local.set.mock.calls[0][0][pageKey];
-      expect(setPayload.notion.pageId).toBe('existing-page-id');
-      expect(setPayload.notion.url).toBe('https://www.notion.so/existing-page-id');
-      expect(setPayload.notion.title).toBe('Updated Title');
-    });
-  });
-
   describe('clearPageState', () => {
     it('應該清除頁面狀態（Phase 3: 刪除 page_* keys）', async () => {
       await service.clearPageState('https://example.com/page');
 
-      // Phase 3: 同時刪除 page_* 新格式 key 和舊格式 saved_* key
-      // Note: computeStableUrl is mocked to return url + '_stable'
       const removeCall = mockStorage.local.remove.mock.calls[0][0];
       expect(removeCall).toContain(`${PAGE_PREFIX}https://example.com/page`);
       expect(removeCall).toContain(`${PAGE_PREFIX}https://example.com/page_stable`);
@@ -994,313 +127,10 @@ describe('StorageService', () => {
     });
   });
 
-  describe('clearNotionState', () => {
-    it('應清除 notion 欄位但保留 highlights', async () => {
-      const pageKey = `${PAGE_PREFIX}https://example.com/page`;
-      mockStorage.local.get.mockResolvedValue({
-        [pageKey]: {
-          highlights: [{ id: '1', text: 'test' }],
-          notion: { pageId: 'abc123', url: 'https://notion.so/abc' },
-          metadata: { lastUpdated: 100 },
-        },
-      });
-
-      await service.clearNotionState('https://example.com/page');
-
-      expect(mockStorage.local.set).toHaveBeenCalledWith(
-        expect.objectContaining({
-          [pageKey]: expect.objectContaining({
-            highlights: [{ id: '1', text: 'test' }],
-            notion: null,
-          }),
-        })
-      );
-    });
-
-    it('page_* 不存在時不應呼叫 set', async () => {
-      mockStorage.local.get.mockResolvedValue({});
-
-      await expect(service.clearNotionState('https://example.com/page')).resolves.toEqual({
-        cleared: true,
-      });
-      expect(mockStorage.local.set).not.toHaveBeenCalled();
-    });
-
-    it('應通過 URL alias 解析並清除 stableUrl 下的 key', async () => {
-      const shortUrl = 'https://example.com/short';
-      const stableUrl = 'https://example.com/stable/';
-      const aliasKey = `${URL_ALIAS_PREFIX}${shortUrl}`;
-      const stablePageKey = `${PAGE_PREFIX}${stableUrl}`;
-
-      mockStorage.local.get
-        .mockResolvedValueOnce({
-          // 第 1 次 get（Phase 4 follow-up: _resolveCanonicalLockKey）：[url_alias:<short>]
-          [aliasKey]: stableUrl,
-        })
-        .mockResolvedValueOnce({
-          // 第 2 次 get（_getPageState）：[page_{short}, saved_{short}, url_alias_{short}]
-          [aliasKey]: stableUrl,
-        })
-        .mockResolvedValueOnce({
-          // 第 3 次 get（_getPageState alias 解析後）：[page_{stable}, saved_{stable}]
-          [stablePageKey]: {
-            highlights: [{ id: '2', text: 'alias test' }],
-            notion: { pageId: 'xyz', url: 'https://notion.so/xyz' },
-            metadata: { lastUpdated: 200 },
-          },
-        });
-
-      await service.clearNotionState(shortUrl);
-
-      // 應寫入 stableUrl 下的 key，而非 shortUrl 的 key
-      expect(mockStorage.local.set).toHaveBeenCalledWith(
-        expect.objectContaining({
-          [stablePageKey]: expect.objectContaining({
-            highlights: [{ id: '2', text: 'alias test' }],
-            notion: null,
-          }),
-        })
-      );
-    });
-
-    it('expectedPageId 不匹配時應跳過清除並記錄 warn', async () => {
-      const pageKey = `${PAGE_PREFIX}https://example.com/page`;
-      mockStorage.local.get.mockResolvedValue({
-        [pageKey]: {
-          highlights: [{ id: '1' }],
-          notion: { pageId: 'correct-id-xyz', url: 'https://notion.so/x' },
-          metadata: { lastUpdated: 100 },
-        },
-      });
-
-      // 傳入不匹配的 expectedPageId
-      await expect(
-        service.clearNotionState('https://example.com/page', {
-          expectedPageId: 'wrong-id-abc',
-        })
-      ).resolves.toEqual({
-        skipped: true,
-        reason: 'pageId_mismatch',
-      });
-
-      // 不應呼叫 set：跳過清除
-      expect(mockStorage.local.set).not.toHaveBeenCalled();
-      expect(mockLogger.warn).toHaveBeenCalledWith(
-        '[StorageService] clearNotionState skipped: pageId mismatch',
-        expect.objectContaining({
-          expectedPageId: 'wron',
-          foundPageId: 'corr',
-        })
-      );
-    });
-
-    it('expectedPageId 匹配時應正常清除 notion 欄位', async () => {
-      const pageKey = `${PAGE_PREFIX}https://example.com/page`;
-      mockStorage.local.get.mockResolvedValue({
-        [pageKey]: {
-          highlights: [{ id: '1' }],
-          notion: { pageId: 'match-id-xyz', url: 'https://notion.so/x' },
-          metadata: { lastUpdated: 100 },
-        },
-      });
-
-      await service.clearNotionState('https://example.com/page', {
-        expectedPageId: 'match-id-xyz',
-      });
-
-      expect(mockStorage.local.set).toHaveBeenCalledWith(
-        expect.objectContaining({
-          [pageKey]: expect.objectContaining({ notion: null }),
-        })
-      );
-    });
-
-    it('未傳 expectedPageId 時行為與舊版相同（無條件清除）', async () => {
-      const pageKey = `${PAGE_PREFIX}https://example.com/page`;
-      mockStorage.local.get.mockResolvedValue({
-        [pageKey]: {
-          highlights: [],
-          notion: { pageId: 'any-id', url: 'https://notion.so/x' },
-          metadata: { lastUpdated: 100 },
-        },
-      });
-
-      // 不傳 options，舊接口相容
-      await service.clearNotionState('https://example.com/page');
-
-      expect(mockStorage.local.set).toHaveBeenCalledWith(
-        expect.objectContaining({
-          [pageKey]: expect.objectContaining({ notion: null }),
-        })
-      );
-    });
-
-    it('clearNotionState 後，original URL 與 stable URL 都不應再回傳 saved data', async () => {
-      const originalUrl = 'https://example.com/original';
-      const stableUrl = 'https://example.com/stable';
-      const pageKey = `${PAGE_PREFIX}${stableUrl}`;
-      const aliasKey = `${URL_ALIAS_PREFIX}${originalUrl}`;
-      const storageData = {
-        [aliasKey]: stableUrl,
-        [pageKey]: {
-          notion: {
-            pageId: 'page-123',
-            url: 'https://www.notion.so/page-123',
-            title: 'Saved page',
-            savedAt: 100,
-            lastVerifiedAt: 100,
-          },
-          highlights: [buildHighlight()],
-          metadata: { createdAt: 100, lastUpdated: 100 },
-        },
-      };
-
-      mockStorage.local.get.mockImplementation(keys => {
-        const keyList = Array.isArray(keys) ? keys : [keys];
-        return Promise.resolve(
-          Object.fromEntries(
-            keyList.filter(key => key in storageData).map(key => [key, storageData[key]])
-          )
-        );
-      });
-      mockStorage.local.set.mockImplementation(async payload => {
-        Object.assign(storageData, payload);
-      });
-      mockStorage.local.remove.mockImplementation(async keys => {
-        const keyList = Array.isArray(keys) ? keys : [keys];
-        keyList.forEach(key => {
-          delete storageData[key];
-        });
-      });
-
-      await service.clearNotionState(originalUrl, {
-        expectedPageId: buildSavedPageData().notionPageId,
-      });
-
-      const originalResult = await service.getSavedPageData(originalUrl);
-      const stableResult = await service.getSavedPageData(stableUrl);
-
-      expect(originalResult).toBeNull();
-      expect(stableResult).toBeNull();
-      expect(storageData[pageKey].highlights).toHaveLength(1);
-      expect(storageData[pageKey].notion).toBeNull();
-    });
-  });
-
-  describe('clearNotionStateWithRetry', () => {
-    it('第一次失敗後應重試一次並回報 recovered', async () => {
-      const clearSpy = jest
-        .spyOn(service, 'clearNotionState')
-        .mockRejectedValueOnce(new Error('temporary storage failure'))
-        .mockResolvedValueOnce();
-
-      const result = await service.clearNotionStateWithRetry('https://example.com/page', {
-        source: 'highlightHandlers',
-        retryDelayMs: 0,
-      });
-
-      expect(clearSpy).toHaveBeenCalledTimes(2);
-      expect(result).toEqual({
-        cleared: true,
-        attempts: 2,
-        recovered: true,
-      });
-      expect(mockLogger.warn).toHaveBeenCalledWith(
-        '[StorageService] clearNotionState 嘗試失敗，準備重試',
-        expect.objectContaining({
-          action: 'clearNotionStateWithRetry',
-          source: 'highlightHandlers',
-          attempt: 1,
-          url: 'https://example.com/page',
-        })
-      );
-      expect(mockLogger.success).toHaveBeenCalledWith(
-        '[StorageService] clearNotionState 重試成功',
-        expect.objectContaining({
-          action: 'clearNotionStateWithRetry',
-          source: 'highlightHandlers',
-          attempts: 2,
-          recovered: true,
-          url: 'https://example.com/page',
-        })
-      );
-    });
-
-    it('兩次嘗試皆失敗時應回傳 cleared: false 並記錄 error 日誌', async () => {
-      const clearSpy = jest
-        .spyOn(service, 'clearNotionState')
-        .mockRejectedValueOnce(new Error('first failure'))
-        .mockRejectedValueOnce(new Error('second failure'));
-
-      const result = await service.clearNotionStateWithRetry('https://example.com/page', {
-        source: 'testSource',
-        retryDelayMs: 0,
-      });
-
-      expect(clearSpy).toHaveBeenCalledTimes(2);
-      expect(result).toEqual({
-        cleared: false,
-        attempts: 2,
-        error: expect.any(Error),
-      });
-      expect(result.error.message).toBe('second failure');
-      expect(mockLogger.warn).toHaveBeenCalledWith(
-        '[StorageService] clearNotionState 嘗試失敗，準備重試',
-        expect.objectContaining({
-          action: 'clearNotionStateWithRetry',
-          source: 'testSource',
-          attempt: 1,
-          error: expect.any(Error),
-        })
-      );
-      expect(mockLogger.error).toHaveBeenCalledWith(
-        '[StorageService] clearNotionState 重試最終失敗',
-        expect.objectContaining({
-          action: 'clearNotionStateWithRetry',
-          source: 'testSource',
-          attempts: 2,
-          recovered: false,
-          error: expect.any(Error),
-        })
-      );
-    });
-
-    it('clearNotionState 回傳 skipped 時應保留 skipped 狀態而非轉成 cleared', async () => {
-      const clearSpy = jest.spyOn(service, 'clearNotionState').mockResolvedValueOnce({
-        skipped: true,
-        reason: 'pageId_mismatch',
-      });
-
-      const result = await service.clearNotionStateWithRetry('https://example.com/page', {
-        source: 'testSource',
-        retryDelayMs: 0,
-        expectedPageId: 'expected-page-id',
-      });
-
-      expect(clearSpy).toHaveBeenCalledTimes(1);
-      expect(result).toEqual({
-        cleared: false,
-        skipped: true,
-        reason: 'pageId_mismatch',
-        attempts: 1,
-        recovered: false,
-      });
-      expect(mockLogger.warn).not.toHaveBeenCalledWith(
-        '[StorageService] clearNotionState 嘗試失敗，準備重試',
-        expect.anything()
-      );
-      expect(mockLogger.error).not.toHaveBeenCalledWith(
-        '[StorageService] clearNotionState 重試最終失敗',
-        expect.anything()
-      );
-    });
-  });
-
   describe('clearLegacyKeys', () => {
     it('應該清除 normalized URL 的三種 keys（Phase 3: page_* + saved_* + highlights_*）', async () => {
       await service.clearLegacyKeys('https://example.com/article?utm_source=test');
 
-      // Phase 3: 刪除 page_*, saved_*, highlights_* 三種前綴
       expect(mockStorage.local.remove).toHaveBeenCalledWith([
         `${PAGE_PREFIX}https://example.com/article`,
         `${SAVED_PREFIX}https://example.com/article`,
@@ -1332,146 +162,6 @@ describe('StorageService', () => {
     });
   });
 
-  describe('savePageDataAndHighlights', () => {
-    const sampleHighlight = {
-      id: 'a1b2c3d4-e5f6-7890-abcd-ef1234567890',
-      text: 'Example highlight text',
-      color: 'yellow',
-      timestamp: 1_700_000_000_000,
-      domPath: 'body > article > p:nth-child(2)',
-    };
-
-    it('應該以 page_* 格式原子寫入頁面數據和標註（Phase 3）', async () => {
-      const pageData = { title: 'Test', pageId: 'page-123' };
-      const highlights = [sampleHighlight];
-      await service.savePageDataAndHighlights('https://example.com/page', pageData, highlights);
-
-      const setCall = mockStorage.local.set.mock.calls[0][0];
-      const pageKey = `${PAGE_PREFIX}https://example.com/page`;
-      expect(setCall[pageKey]).toBeDefined();
-      expect(setCall[pageKey].notion).toBeDefined();
-      expect(setCall[pageKey].highlights).toHaveLength(1);
-      expect(setCall[pageKey].highlights[0].id).toBe(sampleHighlight.id);
-      expect(setCall[pageKey].highlights[0].text).toBe(sampleHighlight.text);
-      expect(setCall[pageKey].highlights[0].color).toBe(sampleHighlight.color);
-      expect(setCall[pageKey].highlights[0].timestamp).toBe(sampleHighlight.timestamp);
-      expect(setCall[pageKey].highlights[0].domPath).toBe(sampleHighlight.domPath);
-      expect(setCall[pageKey].metadata).toBeDefined();
-    });
-
-    it('應該在 pageData 為 null 時仍寫入 highlights', async () => {
-      const highlights = [
-        sampleHighlight,
-        {
-          ...sampleHighlight,
-          id: 'b2c3d4e5-f6a7-8901-bcde-f12345678901',
-          text: 'Second highlight',
-        },
-      ];
-      await service.savePageDataAndHighlights('https://example.com/page', null, highlights);
-
-      const setCall = mockStorage.local.set.mock.calls[0][0];
-      const pageKey = `${PAGE_PREFIX}https://example.com/page`;
-      expect(setCall[pageKey].notion).toBeNull();
-      expect(setCall[pageKey].highlights).toHaveLength(2);
-      expect(setCall[pageKey].highlights[0].id).toBe(sampleHighlight.id);
-    });
-
-    it('應該在 highlights 為 null 時仍寫入 pageData（notion = null 除外）', async () => {
-      const pageData = { title: 'Test', pageId: 'page-123' };
-      await service.savePageDataAndHighlights('https://example.com/page', pageData, null);
-
-      const setCall = mockStorage.local.set.mock.calls[0][0];
-      const pageKey = `${PAGE_PREFIX}https://example.com/page`;
-      expect(setCall[pageKey].notion).toBeDefined();
-      expect(setCall[pageKey].highlights).toEqual([]);
-    });
-
-    it('如果全部為 null 則不應調用 set', async () => {
-      await service.savePageDataAndHighlights('https://example.com/page', null, null);
-      expect(mockStorage.local.set).not.toHaveBeenCalled();
-    });
-
-    it('[CANONICAL] alias 命中時應該寫入 page_<stableUrl> 並清理 page_<original> 與 highlights_<original>', async () => {
-      const originalUrl = 'https://example.com/article?ref=xyz';
-      const stableUrl = 'https://example.com/article';
-      const aliasNormKey = `${URL_ALIAS_PREFIX}${originalUrl}`;
-      const stablePageKey = `${PAGE_PREFIX}${stableUrl}`;
-      const originalPageKey = `${PAGE_PREFIX}${originalUrl}`;
-      const originalLegacyKey = `${HIGHLIGHTS_PREFIX}${originalUrl}`;
-
-      mockStorage.local.get.mockResolvedValue({
-        [aliasNormKey]: stableUrl,
-        [originalPageKey]: { notion: { pageId: 'old' }, highlights: ['stale'], metadata: {} },
-        [originalLegacyKey]: ['legacy-h'],
-      });
-
-      const pageData = { title: 'Article', pageId: 'page-new' };
-      const highlights = [
-        {
-          id: 'a1b2c3d4-e5f6-7890-abcd-ef1234567890',
-          text: 'fresh highlight',
-          color: 'yellow',
-          timestamp: 1_700_000_000_000,
-          domPath: 'body > p',
-        },
-      ];
-
-      await service.savePageDataAndHighlights(originalUrl, pageData, highlights);
-
-      // 寫入應落在 canonical stable page key
-      expect(mockStorage.local.set).toHaveBeenCalledWith({
-        [stablePageKey]: expect.objectContaining({
-          notion: expect.objectContaining({ pageId: 'page-new' }),
-          highlights: expect.arrayContaining([expect.objectContaining({ id: highlights[0].id })]),
-        }),
-      });
-      // MUST NOT 寫入 page_<original>
-      expect(mockStorage.local.set).not.toHaveBeenCalledWith(
-        expect.objectContaining({ [originalPageKey]: expect.anything() })
-      );
-      // legacy keys 必須被清理
-      const removedKeys = mockStorage.local.remove.mock.calls.flatMap(args => args[0]);
-      expect(removedKeys).toEqual(expect.arrayContaining([originalPageKey, originalLegacyKey]));
-    });
-
-    it('legacy cleanup MUST 清理 value 為 null 的歷史毀損 key（key 存在性而非 truthiness）', async () => {
-      const originalUrl = 'https://example.com/article?ref=xyz';
-      const stableUrl = 'https://example.com/article';
-      const aliasNormKey = `${URL_ALIAS_PREFIX}${originalUrl}`;
-      const stablePageKey = `${PAGE_PREFIX}${stableUrl}`;
-      const originalPageKey = `${PAGE_PREFIX}${originalUrl}`;
-
-      // 模擬歷史毀損:page_<original> 在 storage 中存在,但 value 是 null
-      // (例如舊版 bug 寫過 null、或半途失敗的 set)
-      mockStorage.local.get.mockResolvedValue({
-        [aliasNormKey]: stableUrl,
-        [originalPageKey]: null,
-      });
-
-      const pageData = { title: 'Article', pageId: 'page-new' };
-      const highlights = [
-        {
-          id: 'a1b2c3d4-e5f6-7890-abcd-ef1234567890',
-          text: 'fresh highlight',
-          color: 'yellow',
-          timestamp: 1_700_000_000_000,
-          domPath: 'body > p',
-        },
-      ];
-
-      await service.savePageDataAndHighlights(originalUrl, pageData, highlights);
-
-      // 寫入應落在 canonical stable page key
-      expect(mockStorage.local.set).toHaveBeenCalledWith({
-        [stablePageKey]: expect.any(Object),
-      });
-      // 即使 value 為 null,毀損 key 仍 MUST 被 remove(語義是「key 存在於 storage」)
-      const removedKeys = mockStorage.local.remove.mock.calls.flatMap(args => args[0]);
-      expect(removedKeys).toEqual(expect.arrayContaining([originalPageKey]));
-    });
-  });
-
   describe('setUrlAlias', () => {
     it('應該設定 URL alias 映射', async () => {
       const originalUrl = 'https://example.com/original';
@@ -1492,7 +182,6 @@ describe('StorageService', () => {
     it('如果 URL 標準化後相同則不應設定 alias', async () => {
       const url1 = 'https://example.com/page?utm_source=test';
       const url2 = 'https://example.com/page';
-      // normalizeUrl is the real implementation, so these URLs become identical, preventing alias creation
       await service.setUrlAlias(url1, url2);
       expect(mockStorage.local.set).not.toHaveBeenCalled();
     });
@@ -1540,405 +229,6 @@ describe('StorageService', () => {
         notionDataSourceId: 'ds_local_123',
       });
       expect(mockStorage.sync.set).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('updateHighlights', () => {
-    it('應該更新 page_* 的 highlights 欄位（Phase 3）', async () => {
-      const existingPageData = {
-        notion: { pageId: 'p123' },
-        highlights: ['old-h1'],
-        metadata: { createdAt: 1000, lastUpdated: 1000 },
-      };
-      mockStorage.local.get.mockResolvedValue({
-        [`${PAGE_PREFIX}https://example.com/page`]: existingPageData,
-      });
-
-      const newHighlights = ['h1', 'h2'];
-      await service.updateHighlights('https://example.com/page', newHighlights);
-
-      expect(mockStorage.local.set).toHaveBeenCalledWith({
-        [`${PAGE_PREFIX}https://example.com/page`]: expect.objectContaining({
-          notion: { pageId: 'p123' },
-          highlights: newHighlights,
-          metadata: expect.objectContaining({ lastUpdated: expect.any(Number) }),
-        }),
-      });
-    });
-
-    it('如果沒有現有 page_* 資料，應該創建新的 page_* 結構（notion 為 null）', async () => {
-      mockStorage.local.get.mockResolvedValue({});
-
-      const newHighlights = ['h1'];
-      await service.updateHighlights('https://example.com/page', newHighlights);
-
-      expect(mockStorage.local.set).toHaveBeenCalledWith({
-        [`${PAGE_PREFIX}https://example.com/page`]: expect.objectContaining({
-          notion: null,
-          highlights: newHighlights,
-        }),
-      });
-    });
-
-    it('如果有舊版 highlights_* 資料，應該升級到 page_* 格式並刪除舊 key', async () => {
-      mockStorage.local.get.mockResolvedValue({
-        [`${HIGHLIGHTS_PREFIX}https://example.com/page`]: ['old_h1'],
-      });
-
-      const newHighlights = ['h1'];
-      await service.updateHighlights('https://example.com/page', newHighlights);
-
-      expect(mockStorage.local.set).toHaveBeenCalledWith({
-        [`${PAGE_PREFIX}https://example.com/page`]: expect.objectContaining({
-          notion: null,
-          highlights: newHighlights,
-        }),
-      });
-      // 舊 key 應被非同步刪除（非同步，不一定即時呼叫）
-    });
-
-    it('應該在刪除舊 highlights_* key 失敗時記錄 debug 日誌', async () => {
-      mockStorage.local.get.mockResolvedValue({
-        [`${HIGHLIGHTS_PREFIX}https://example.com/page`]: ['old_h1'],
-      });
-      mockStorage.local.remove.mockRejectedValue(new Error('Remove highlights failed'));
-      mockLogger.debug = jest.fn();
-
-      await service.updateHighlights('https://example.com/page', ['h1']);
-
-      // 確保 set 已經被呼叫
-      expect(mockStorage.local.set).toHaveBeenCalled();
-
-      // 等待微任務佇列清空，讓 .catch 執行
-      await new Promise(process.nextTick);
-
-      expect(mockLogger.debug).toHaveBeenCalledWith(
-        '[StorageService] Failed to remove legacy keys',
-        expect.objectContaining({
-          error: 'Remove highlights failed',
-          keys: expect.arrayContaining([`${HIGHLIGHTS_PREFIX}https://example.com/page`]),
-        })
-      );
-    });
-
-    it('應該在 storage.local.set 失敗時記錄錯誤並拋出（寫入階段）', async () => {
-      mockStorage.local.get.mockResolvedValue({
-        [`${PAGE_PREFIX}https://example.com/page`]: {
-          notion: null,
-          highlights: ['old-h1'],
-          metadata: {},
-        },
-      });
-
-      mockStorage.local.set.mockRejectedValue(new Error('Write failed'));
-
-      await expect(service.updateHighlights('https://example.com/page', ['h1'])).rejects.toThrow(
-        'Write failed'
-      );
-
-      expect(mockLogger.error).toHaveBeenCalledWith(
-        '[StorageService] updateHighlights failed',
-        expect.objectContaining({ error: expect.any(Error) })
-      );
-    });
-
-    // === Step 1: Contract-driven mutation tests（2026-05-03 completion plan） ===
-
-    it('alias 命中時應該寫入 page_<stableUrl> 並 cleanup legacy keys', async () => {
-      const originalUrl = 'https://example.com/article?ref=abc';
-      const stableUrl = 'https://example.com/article';
-
-      const aliasNormKey = `${URL_ALIAS_PREFIX}${originalUrl}`;
-      const stablePageKey = `${PAGE_PREFIX}${stableUrl}`;
-      const originalPageKey = `${PAGE_PREFIX}${originalUrl}`;
-      const originalLegacyKey = `${HIGHLIGHTS_PREFIX}${originalUrl}`;
-      const stableLegacyKey = `${HIGHLIGHTS_PREFIX}${stableUrl}`;
-
-      // Storage 內容：
-      // - alias: original → stable
-      // - page_<stable> 已存在（之前 alias 命中過）
-      // - 殘留 page_<original> 與 highlights_<original>（過去未清乾淨）
-      mockStorage.local.get.mockResolvedValue({
-        [aliasNormKey]: stableUrl,
-        [stablePageKey]: {
-          notion: { pageId: 'stable-pid' },
-          highlights: ['old-stable-h1'],
-          metadata: { createdAt: 1000, lastUpdated: 1500 },
-        },
-        [originalPageKey]: {
-          notion: null,
-          highlights: ['old-original-h1'],
-          metadata: { createdAt: 800, lastUpdated: 900 },
-        },
-        [originalLegacyKey]: ['old-legacy-h1'],
-      });
-
-      const newHighlights = ['h1', 'h2'];
-      await service.updateHighlights(originalUrl, newHighlights);
-
-      // 寫入應該落在 page_<stable> canonical key，且保留 stable 的 notion 欄位
-      expect(mockStorage.local.set).toHaveBeenCalledWith({
-        [stablePageKey]: expect.objectContaining({
-          notion: { pageId: 'stable-pid' },
-          highlights: newHighlights,
-          metadata: expect.objectContaining({ lastUpdated: expect.any(Number) }),
-        }),
-      });
-      // MUST NOT 寫入 page_<original>
-      expect(mockStorage.local.set).not.toHaveBeenCalledWith(
-        expect.objectContaining({ [originalPageKey]: expect.anything() })
-      );
-
-      // 等待非同步 cleanup
-      await new Promise(process.nextTick);
-
-      // legacy keys 應被清除（含 page_<original> 與 highlights_<original>）
-      const removeCalls = mockStorage.local.remove.mock.calls.map(args => args[0]);
-      const removedKeys = removeCalls.flat();
-      expect(removedKeys).toEqual(expect.arrayContaining([originalPageKey, originalLegacyKey]));
-      // MUST NOT remove canonical key 自身
-      expect(removedKeys).not.toContain(stablePageKey);
-      // 不存在的 stableLegacyKey 不應出現在 remove 名單（contract 過濾掉）
-      expect(removedKeys).not.toContain(stableLegacyKey);
-    });
-
-    it('並發 update：以 stable / original 兩個 pageUrl 同時呼叫應序列化於同一 canonical lock', async () => {
-      const originalUrl = 'https://example.com/post?utm=x';
-      const stableUrl = 'https://example.com/post';
-
-      const aliasNormKey = `${URL_ALIAS_PREFIX}${originalUrl}`;
-      const stablePageKey = `${PAGE_PREFIX}${stableUrl}`;
-
-      // 模擬 alias 命中、且 page_<stable> 初始為空（兩次呼叫都會建立/覆寫 page_<stable>）
-      // 為驗證互鎖：mockStorage.local.set 被 instrumented 為記錄 set 順序與 in-flight count。
-      let setInFlight = 0;
-      let maxConcurrent = 0;
-      const setOrder = [];
-
-      mockStorage.local.get.mockImplementation(async keys => {
-        // 模擬非零延遲，放大 race window
-        await new Promise(resolve => setTimeout(resolve, 5));
-        const all = {
-          [aliasNormKey]: stableUrl,
-          [`${URL_ALIAS_PREFIX}${stableUrl}`]: stableUrl,
-        };
-        // 動態回傳 set 後的 stable page state（用 closure 模擬）
-        if (setOrder.length > 0) {
-          // 上一次寫入的 highlights
-          all[stablePageKey] = {
-            notion: null,
-            highlights: setOrder.at(-1),
-            metadata: { createdAt: 1, lastUpdated: 2 },
-          };
-        }
-        const requested = Array.isArray(keys) ? keys : [keys];
-        const acc = {};
-        for (const k of requested) {
-          if (k in all) {
-            acc[k] = all[k];
-          }
-        }
-        return acc;
-      });
-
-      mockStorage.local.set.mockImplementation(async payload => {
-        setInFlight++;
-        maxConcurrent = Math.max(maxConcurrent, setInFlight);
-        await new Promise(resolve => setTimeout(resolve, 5));
-        // 紀錄寫入的 highlights
-        const value = payload[stablePageKey];
-        setOrder.push(value.highlights);
-        setInFlight--;
-      });
-
-      // 同時以 stable 與 original 兩個 URL 啟動 update
-      await Promise.all([
-        service.updateHighlights(stableUrl, ['from-stable']),
-        service.updateHighlights(originalUrl, ['from-original']),
-      ]);
-
-      // 互鎖驗證：兩次 set 不應同時 in-flight（max concurrent === 1）
-      expect(maxConcurrent).toBe(1);
-      // 兩次呼叫都寫入 canonical key（page_<stable>），不會落到 page_<original>
-      expect(setOrder).toHaveLength(2);
-      const setKeys = mockStorage.local.set.mock.calls.map(c => Object.keys(c[0])[0]);
-      expect(setKeys.every(k => k === stablePageKey)).toBe(true);
-    });
-
-    it('[CANONICAL] legacy cleanup MUST 在 updateHighlights 回傳前完成（不依賴 process.nextTick）', async () => {
-      const originalUrl = 'https://example.com/article?ref=abc';
-      const stableUrl = 'https://example.com/article';
-      const aliasNormKey = `${URL_ALIAS_PREFIX}${originalUrl}`;
-      const stablePageKey = `${PAGE_PREFIX}${stableUrl}`;
-      const originalPageKey = `${PAGE_PREFIX}${originalUrl}`;
-      const originalLegacyKey = `${HIGHLIGHTS_PREFIX}${originalUrl}`;
-
-      mockStorage.local.get.mockResolvedValue({
-        [aliasNormKey]: stableUrl,
-        [stablePageKey]: { notion: { pageId: 'p1' }, highlights: [], metadata: {} },
-        [originalPageKey]: { notion: null, highlights: ['old'], metadata: {} },
-        [originalLegacyKey]: ['legacy-h'],
-      });
-
-      // 阻塞 mockStorage.local.remove 直到外部解 promise；用以驗證 lock-scope 同步性。
-      // 修復前（fire-and-forget）：updateHighlights 已 resolve，updateResolved === true。
-      // 修復後（await + try/catch）：updateHighlights 仍 pending 直到 cleanup 結束。
-      let resolveRemove;
-      const removePending = new Promise(resolve => {
-        resolveRemove = resolve;
-      });
-      mockStorage.local.remove.mockImplementation(() => removePending);
-
-      let updateResolved = false;
-      const updatePromise = service.updateHighlights(originalUrl, ['h1']).then(() => {
-        updateResolved = true;
-      });
-
-      // 給 microtask queue 足夠時間執行所有 await（set / get），但 cleanup 仍被阻塞。
-      await new Promise(resolve => setTimeout(resolve, 10));
-
-      expect(updateResolved).toBe(false);
-
-      // 解開 remove → cleanup 完成 → lock 釋放 → updateHighlights resolve
-      resolveRemove();
-      await updatePromise;
-      expect(updateResolved).toBe(true);
-
-      const removedKeys = mockStorage.local.remove.mock.calls.flatMap(args => args[0]);
-      expect(removedKeys).toEqual(expect.arrayContaining([originalPageKey, originalLegacyKey]));
-    });
-  });
-
-  describe('Canonical lock alignment across writers', () => {
-    // Follow-up plan 2026-05-03-highlight-canonical-lock-and-delete-all-cleanup-followup §1：
-    // 所有 page-state writers 在 alias 命中時 MUST 與 updateHighlights 共用同一條
-    // canonical lock 推導規則（contract.mutationTargetKey），不再各自實作。
-    it('updateHighlights 與 setSavedPageData 在 alias 命中時 MUST 鎖在同一個 page_<stable>', async () => {
-      const originalUrl = 'https://example.com/post?utm=x';
-      const stableUrl = 'https://example.com/post';
-      const stablePageKey = `${PAGE_PREFIX}${stableUrl}`;
-
-      makeStorageMock({
-        [`${URL_ALIAS_PREFIX}${originalUrl}`]: stableUrl,
-        [`${URL_ALIAS_PREFIX}${stableUrl}`]: stableUrl,
-        [stablePageKey]: {
-          notion: null,
-          highlights: [{ id: 'seed', text: 'seed', color: 'yellow' }],
-          metadata: { lastUpdated: 1 },
-        },
-      });
-
-      const lockSpy = jest.spyOn(service, '_withLock');
-
-      await Promise.all([
-        service.updateHighlights(originalUrl, [{ id: 'h-update', text: 'u', color: 'yellow' }]),
-        service.setSavedPageData(originalUrl, {
-          notionPageId: 'page-1',
-          notionUrl: 'https://notion.so/page-1',
-          title: 'T',
-        }),
-      ]);
-
-      const lockKeys = lockSpy.mock.calls.map(call => call[0]);
-      expect(lockKeys).toHaveLength(2);
-      expect(lockKeys.every(k => k === stablePageKey)).toBe(true);
-    });
-
-    it('savePageDataAndHighlights 在 alias 命中時 MUST 鎖在 page_<stable>', async () => {
-      const originalUrl = 'https://example.com/blog?ref=x';
-      const stableUrl = 'https://example.com/blog';
-      const stablePageKey = `${PAGE_PREFIX}${stableUrl}`;
-
-      makeStorageMock({
-        [`${URL_ALIAS_PREFIX}${originalUrl}`]: stableUrl,
-        [`${URL_ALIAS_PREFIX}${stableUrl}`]: stableUrl,
-        [stablePageKey]: {
-          notion: null,
-          highlights: [],
-          metadata: { lastUpdated: 1 },
-        },
-      });
-
-      const lockSpy = jest.spyOn(service, '_withLock');
-
-      await service.savePageDataAndHighlights(originalUrl, { title: 'T' }, [
-        { id: 'h', text: 't', color: 'yellow' },
-      ]);
-
-      const lockKeys = lockSpy.mock.calls.map(call => call[0]);
-      expect(lockKeys[0]).toBe(stablePageKey);
-    });
-
-    it('clearNotionState 在 alias 命中時 MUST 鎖在 page_<stable>', async () => {
-      const originalUrl = 'https://example.com/article?utm=ig';
-      const stableUrl = 'https://example.com/article';
-      const stablePageKey = `${PAGE_PREFIX}${stableUrl}`;
-
-      makeStorageMock({
-        [`${URL_ALIAS_PREFIX}${originalUrl}`]: stableUrl,
-        [`${URL_ALIAS_PREFIX}${stableUrl}`]: stableUrl,
-        [stablePageKey]: {
-          notion: { pageId: 'pid', url: 'https://notion.so/pid' },
-          highlights: [{ id: 'h', text: 't', color: 'yellow' }],
-          metadata: { lastUpdated: 1 },
-        },
-      });
-
-      const lockSpy = jest.spyOn(service, '_withLock');
-
-      await service.clearNotionState(originalUrl);
-
-      const lockKeys = lockSpy.mock.calls.map(call => call[0]);
-      expect(lockKeys[0]).toBe(stablePageKey);
-    });
-
-    // Lite follow-up（2026-05-04）：剩餘兩個 _withLock writers 也對齊到 canonical helper
-    it('removeSavedPageData 在 alias 命中時 MUST 鎖在 page_<stable>', async () => {
-      const originalUrl = 'https://example.com/news?ref=fb';
-      const stableUrl = 'https://example.com/news';
-      const stablePageKey = `${PAGE_PREFIX}${stableUrl}`;
-
-      makeStorageMock({
-        [`${URL_ALIAS_PREFIX}${originalUrl}`]: stableUrl,
-        [`${URL_ALIAS_PREFIX}${stableUrl}`]: stableUrl,
-        [stablePageKey]: {
-          notion: { pageId: 'pid', url: 'https://notion.so/pid' },
-          highlights: [{ id: 'h', text: 'k', color: 'yellow' }],
-          metadata: { lastUpdated: 1 },
-        },
-      });
-
-      const lockSpy = jest.spyOn(service, '_withLock');
-
-      await service.removeSavedPageData(originalUrl);
-
-      const lockKeys = lockSpy.mock.calls.map(call => call[0]);
-      expect(lockKeys[0]).toBe(stablePageKey);
-    });
-
-    it('_triggerReadTimeUpgrade 在 alias 命中時 MUST 鎖在 page_<stable>', async () => {
-      const originalUrl = 'https://example.com/recipe?utm=tw';
-      const stableUrl = 'https://example.com/recipe';
-      const stablePageKey = `${PAGE_PREFIX}${stableUrl}`;
-      const savedKey = `${SAVED_PREFIX}${originalUrl}`;
-      const savedData = { notionPageId: 'pid', title: 'Recipe' };
-
-      makeStorageMock({
-        [`${URL_ALIAS_PREFIX}${originalUrl}`]: stableUrl,
-        [`${URL_ALIAS_PREFIX}${stableUrl}`]: stableUrl,
-      });
-
-      const lockSpy = jest.spyOn(service, '_withLock');
-
-      // _triggerReadTimeUpgrade 為 fire-and-forget,await 確保 lock pre-resolve 完成
-      await service._triggerReadTimeUpgrade(originalUrl, savedData, savedKey);
-      // 等待 fire-and-forget _withLock 微任務排隊完成
-      await Promise.resolve();
-      await Promise.resolve();
-
-      const lockKeys = lockSpy.mock.calls.map(call => call[0]);
-      expect(lockKeys.length).toBeGreaterThan(0);
-      expect(lockKeys[0]).toBe(stablePageKey);
     });
   });
